requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
drf-yasg>=1.21.0
django-filter>=23.0 